+++ test/IntegrationTest.t.sol
@@ -28,19 +28,14 @@ contract IntegrationTest is SetupHook {
 
   function testCompleteSwapAndFillWorkflow() public {
     uint256 swapAmount = 2 ether;
-    
+
     // Alice creates async swap order
     vm.startPrank(alice);
     token0.approve(address(router), swapAmount);
-    
-    AsyncOrder memory aliceOrder = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: swapAmount,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory aliceOrder =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: swapAmount, sqrtPrice: 2 ** 96 });
+
     router.swap(aliceOrder, abi.encode(alice, address(router)));
     vm.stopPrank();
 
@@ -62,34 +57,24 @@ contract IntegrationTest is SetupHook {
   function testMultipleUsersMultipleOrders() public {
     uint256 aliceAmount = 1 ether;
     uint256 bobAmount = 1.5 ether;
-    
+
     // Alice creates order (zeroForOne = true)
     vm.startPrank(alice);
     token0.approve(address(router), aliceAmount);
-    
-    AsyncOrder memory aliceOrder = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: aliceAmount,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory aliceOrder =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: aliceAmount, sqrtPrice: 2 ** 96 });
+
     router.swap(aliceOrder, abi.encode(alice, address(router)));
     vm.stopPrank();
 
     // Bob creates order (zeroForOne = false)
     vm.startPrank(bob);
     token1.approve(address(router), bobAmount);
-    
-    AsyncOrder memory bobOrder = AsyncOrder({
-      key: key,
-      owner: bob,
-      zeroForOne: false,
-      amountIn: bobAmount,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory bobOrder =
+      AsyncOrder({ key: key, owner: bob, zeroForOne: false, amountIn: bobAmount, sqrtPrice: 2 ** 96 });
+
     router.swap(bobOrder, abi.encode(bob, address(router)));
     vm.stopPrank();
 
@@ -106,15 +91,10 @@ contract IntegrationTest is SetupHook {
     // Charlie fills Bob's order
     vm.startPrank(charlie);
     token0.approve(address(router), bobAmount);
-    
-    AsyncOrder memory bobFillOrder = AsyncOrder({
-      key: key,
-      owner: bob,
-      zeroForOne: false,
-      amountIn: bobAmount,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory bobFillOrder =
+      AsyncOrder({ key: key, owner: bob, zeroForOne: false, amountIn: bobAmount, sqrtPrice: 2 ** 96 });
+
     router.fillOrder(bobFillOrder, "");
     vm.stopPrank();
 
@@ -129,33 +109,23 @@ contract IntegrationTest is SetupHook {
     uint256 secondSwap = 2 ether;
     uint256 firstFill = 1.5 ether;
     uint256 secondFill = 1.5 ether;
-    
+
     // Alice creates first order
     vm.startPrank(alice);
     token0.approve(address(router), firstSwap);
-    
-    AsyncOrder memory aliceOrder1 = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: firstSwap,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory aliceOrder1 =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: firstSwap, sqrtPrice: 2 ** 96 });
+
     router.swap(aliceOrder1, abi.encode(alice, address(router)));
     assertEq(hook.asyncOrder(poolId, alice, true), firstSwap);
 
     // Alice creates second order (accumulates)
     token0.approve(address(router), secondSwap);
-    
-    AsyncOrder memory aliceOrder2 = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: secondSwap,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory aliceOrder2 =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: secondSwap, sqrtPrice: 2 ** 96 });
+
     router.swap(aliceOrder2, abi.encode(alice, address(router)));
     vm.stopPrank();
 
@@ -165,15 +135,10 @@ contract IntegrationTest is SetupHook {
     // Bob partially fills
     vm.startPrank(bob);
     token1.approve(address(router), firstFill);
-    
-    AsyncOrder memory fillOrder1 = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: firstFill,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory fillOrder1 =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: firstFill, sqrtPrice: 2 ** 96 });
+
     router.fillOrder(fillOrder1, "");
     vm.stopPrank();
 
@@ -183,15 +148,10 @@ contract IntegrationTest is SetupHook {
     // Charlie fills remainder
     vm.startPrank(charlie);
     token1.approve(address(router), secondFill);
-    
-    AsyncOrder memory fillOrder2 = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: secondFill,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory fillOrder2 =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: secondFill, sqrtPrice: 2 ** 96 });
+
     router.fillOrder(fillOrder2, "");
     vm.stopPrank();
 
@@ -202,20 +162,15 @@ contract IntegrationTest is SetupHook {
   function testBatchOrderExecution() public {
     uint256 orderCount = 5;
     uint256 amountPerOrder = 0.5 ether;
-    
+
     // Alice creates multiple orders
     vm.startPrank(alice);
-    for (uint i = 0; i < orderCount; i++) {
+    for (uint256 i = 0; i < orderCount; i++) {
       token0.approve(address(router), amountPerOrder);
-      
-      AsyncOrder memory order = AsyncOrder({
-        key: key,
-        owner: alice,
-        zeroForOne: true,
-        amountIn: amountPerOrder,
-        sqrtPrice: 2 ** 96
-      });
-      
+
+      AsyncOrder memory order =
+        AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: amountPerOrder, sqrtPrice: 2 ** 96 });
+
       router.swap(order, abi.encode(alice, address(router)));
     }
     vm.stopPrank();
@@ -224,14 +179,9 @@ contract IntegrationTest is SetupHook {
     assertEq(hook.asyncOrder(poolId, alice, true), orderCount * amountPerOrder);
 
     // Execute orders one by one using router (since alice set router as executor)
-    for (uint i = 0; i < orderCount; i++) {
-      AsyncOrder memory fillOrder = AsyncOrder({
-        key: key,
-        owner: alice,
-        zeroForOne: true,
-        amountIn: amountPerOrder,
-        sqrtPrice: 2 ** 96
-      });
+    for (uint256 i = 0; i < orderCount; i++) {
+      AsyncOrder memory fillOrder =
+        AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: amountPerOrder, sqrtPrice: 2 ** 96 });
 
       vm.startPrank(bob);
       token1.approve(address(router), amountPerOrder);
@@ -245,48 +195,38 @@ contract IntegrationTest is SetupHook {
 
   function testCrossDirectionalOrders() public {
     uint256 amount = 1 ether;
-    
+
     // Alice: token0 -> token1 (zeroForOne = true)
     vm.startPrank(alice);
     token0.approve(address(router), amount);
-    
-    AsyncOrder memory aliceOrder = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: amount,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory aliceOrder =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: amount, sqrtPrice: 2 ** 96 });
+
     router.swap(aliceOrder, abi.encode(alice, address(router)));
     vm.stopPrank();
 
     // Bob: token1 -> token0 (zeroForOne = false)
     vm.startPrank(bob);
     token1.approve(address(router), amount);
-    
-    AsyncOrder memory bobOrder = AsyncOrder({
-      key: key,
-      owner: bob,
-      zeroForOne: false,
-      amountIn: amount,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory bobOrder =
+      AsyncOrder({ key: key, owner: bob, zeroForOne: false, amountIn: amount, sqrtPrice: 2 ** 96 });
+
     router.swap(bobOrder, abi.encode(bob, address(router)));
     vm.stopPrank();
 
     // Charlie can fill both orders
     vm.startPrank(charlie);
-    
+
     // Fill Alice's order (needs token1)
     token1.approve(address(router), amount);
     router.fillOrder(aliceOrder, "");
-    
-    // Fill Bob's order (needs token0) 
+
+    // Fill Bob's order (needs token0)
     token0.approve(address(router), amount);
     router.fillOrder(bobOrder, "");
-    
+
     vm.stopPrank();
 
     // Verify both filled
@@ -296,23 +236,18 @@ contract IntegrationTest is SetupHook {
 
   function testAlgorithmIntegration() public {
     uint256 amount = 1 ether;
-    
+
     // Verify algorithm is set correctly
     address algorithmAddress = address(hook.ALGORITHM());
     assertTrue(algorithmAddress != address(0));
-    
+
     // Create order to trigger algorithm
     vm.startPrank(alice);
     token0.approve(address(router), amount);
-    
-    AsyncOrder memory order = AsyncOrder({
-      key: key,
-      owner: alice,
-      zeroForOne: true,
-      amountIn: amount,
-      sqrtPrice: 2 ** 96
-    });
-    
+
+    AsyncOrder memory order =
+      AsyncOrder({ key: key, owner: alice, zeroForOne: true, amountIn: amount, sqrtPrice: 2 ** 96 });
+
     router.swap(order, abi.encode(alice, address(router)));
     vm.stopPrank();
 
@@ -323,26 +258,21 @@ contract IntegrationTest is SetupHook {
   function testLargeVolumeStressTest() public {
     uint256 userCount = 10;
     uint256 amountPerUser = 0.1 ether;
-    
+
     // Multiple users create orders
-    for (uint i = 0; i < userCount; i++) {
+    for (uint256 i = 0; i < userCount; i++) {
       address user = address(uint160(i + 1000)); // Generate unique addresses
       topUp(user, amountPerUser);
-      
+
       vm.startPrank(user);
       token0.approve(address(router), amountPerUser);
-      
-      AsyncOrder memory order = AsyncOrder({
-        key: key,
-        owner: user,
-        zeroForOne: true,
-        amountIn: amountPerUser,
-        sqrtPrice: 2 ** 96
-      });
-      
+
+      AsyncOrder memory order =
+        AsyncOrder({ key: key, owner: user, zeroForOne: true, amountIn: amountPerUser, sqrtPrice: 2 ** 96 });
+
       router.swap(order, abi.encode(user, address(router)));
       vm.stopPrank();
-      
+
       // Verify each order
       assertEq(hook.asyncOrder(poolId, user, true), amountPerUser);
     }
@@ -350,18 +280,13 @@ contract IntegrationTest is SetupHook {
     // Single filler fills all orders
     vm.startPrank(charlie);
     token1.approve(address(router), userCount * amountPerUser);
-    
-    for (uint i = 0; i < userCount; i++) {
+
+    for (uint256 i = 0; i < userCount; i++) {
       address user = address(uint160(i + 1000));
-      
-      AsyncOrder memory fillOrder = AsyncOrder({
-        key: key,
-        owner: user,
-        zeroForOne: true,
-        amountIn: amountPerUser,
-        sqrtPrice: 2 ** 96
-      });
-      
+
+      AsyncOrder memory fillOrder =
+        AsyncOrder({ key: key, owner: user, zeroForOne: true, amountIn: amountPerUser, sqrtPrice: 2 ** 96 });
+
       router.fillOrder(fillOrder, "");
       assertEq(hook.asyncOrder(poolId, user, true), 0);
     }
